Adiciona handle_spy v1.

+++ src/processaQry.h
@@ -5,6 +5,6 @@
 #include "Lista.h"
 
 SmuTreap processa_qry(SmuTreap t, const char *pathQry, const char *pathSaida, const char *nomeQry, Lista *array_anotacoes, 
-                        Lista lista_anotacoes_svg, int *idMax, int epsilon);
+                        Lista lista_anotacoes_svg, int *idMax, double epsilon);
 
 #endif // PROCESSA_QRY_H
